Update
Various clarity for javadoc comments and formatting

IEnergyHandler now extends IEnergyProvider and IEnergyReceiver for the
purpose of enets making optimizations.
It is not required that machines change anything but it is recommended
for clarity. Enets should update at the earliest convenience to account
for machines that will begin having finer-grained implementations.